+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ## v0.2
 
 ### v0.2.8
+- 🔶 deprecate `repeat` renamed to `cycle`
 - 🔶 deprecate `loop` renamed to `count`
 - 🔶 deprecate `loop2d` renamed to `count2d`
 - 🍏 add `collect` fn